+++ UORespawnApp/Scripts/Helpers/Point2D.cs
@@ -1,0 +1,8 @@
+namespace UORespawnApp.Scripts.Helpers
+{
+    internal struct Point2D(int x, int y)
+    {
+        public int X { get; set; } = x;
+        public int Y { get; set; } = y;
+    }
+}