+++ UORespawnApp/Scripts/Helpers/Point3D.cs
@@ -1,9 +1,9 @@
 ﻿namespace UORespawnApp.Scripts.Helpers
 {
-    internal struct Point3D
+    internal struct Point3D(int x, int y, int z)
     {
-        public int X { get; set; }
-        public int Y { get; set; }
-        public int Z { get; set; }
+        public int X { get; set; } = x;
+        public int Y { get; set; } = y;
+        public int Z { get; set; } = z;
     }
 }